Add a first version of a new tab 'config.xml utility'
Will have controls to work with the config.xml more directly and easier than
it's done in the current page that only shows a list of links.
The entire skeleton for this new tab was added, but nothing was really implemented
yet in the javascript.
Includes a lot of changes especially in the views to organize them better.

refs #1110

Conflicts:
	src/api_mate.jade