Reorganize and make the code more modular
Specially the js class, that is now more independent from the html markup than
it was before. It now expects certain "data" attributes in the html to find the
elements it needs, instead of expecting ids.
The goal is to have a js class that can be used in other environments and not
only in this application. This is the first step toward this goal.