Remove remaining tooltips when we update URLs
The tooltip freezes everytime we update URLs during a visualization (for example, 'tab-ing' beetwen inputs while mouse is over a tooltip). This happens because tooltip's mark is lost when we update URLs, once we rewrite all placeholder's inner html.  Since tooltip is appended to the body tag (and not as a child of placeholder), it stays in the page when we replace the code with new URLs.